% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnet_concatenate.R
\name{mnet_concatenate}
\alias{mnet_concatenate}
\title{Concatenate subdaily Oklahoma Mesonet records for multiple dates or stations.}
\usage{
mnet_concatenate(
  stid = NULL,
  start_date = NULL,
  end_date = NULL,
  site_info = NULL,
  file_cache = NULL
)
}
\arguments{
\item{stid}{a character vector of four-digit station identifiers
for the Mesonet stations from which to download data}

\item{start_date}{the first date for which to download data specified in
Central Standard Time (i.e. "America/Costa_Rica") zone as a Date or POSIXt
object or a string with the format YYYY-MM-DD, where YYYY is the four-digit
year, MM is the two-digit numeric month and DD is the two-digit day of month}

\item{end_date}{the final date for which to download data specified in
Central Standard Time (i.e. "America/Costa_Rica") zone as a Date or POSIXt
object or a string with the format YYYY-MM-DD, where YYYY is the four-digit
year, MM is the two-digit numeric month and DD is the two-digit day of month}

\item{site_info}{a data frame with site and soil information for each
Oklahoma Mesonet station as returned by \link[mesonet]{mnet_site_info}. If
NULL, \link[mesonet]{mnet_site_info} will be used to download this
information internally}

\item{file_cache}{a character string providing a path to the local Mesonet
file cache. If NULL, the function will search for the local file cache and
if not found will prompt the user to create one.}
}
\description{
Concatenate subdaily Oklahoma Mesonet records for multiple dates or stations.
}
