% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnet_drop_units.R
\name{mnet_drop_units}
\alias{mnet_drop_units}
\title{Drop units from all columns of Oklahoma Mesonet data frame}
\usage{
mnet_drop_units(df)
}
\arguments{
\item{df}{a data frame with units columns}
}
\value{
A data frame with identical data, but without units
}
\description{
Drop units from all columns of Oklahoma Mesonet data frame
}
