% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnet_subdaily_c.R
\name{mnet_subdaily_c}
\alias{mnet_subdaily_c}
\title{Concatenate subdaily Oklahoma Mesonet records}
\usage{
mnet_subdaily_c(file_cache, mc.cores = 1)
}
\arguments{
\item{file_cache}{a character string providing a path to the local Mesonet
file cache. If NULL, the function will search for the local file cache and
if not found will prompt the user to create one.}
}
\description{
Concatenate subdaily Oklahoma Mesonet records for multiple dates or stations.
Rainfall
}
