% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnet_variable_definition.R
\name{mnet_variable_definition}
\alias{mnet_variable_definition}
\title{Find Mesonet variable definition}
\usage{
mnet_variable_definition(
  id,
  columns = c("ID", "Variable Name", "Unit", "Description")
)
}
\arguments{
\item{id}{a string or \link[base]{regular expression} to use for retrieving definitions
of Mesonet variables}

\item{columns}{a character vector of columns to include in the
output. Possible values include any combination of "ID", "Variable Name",
"Unit" or "Description"}
}
\value{
a data frame with Mesonet variable definitions filtered by the \code{id}
 argument
}
\description{
Provide definition of a requested set of Mesonet variables including (by
 default) the standard variable identifier (ID), name, unit, and description.
}
\examples{

mnet_variable_definition("RAIN")

mnet_variable_definition("^2")

mnet_variable_definition("^2",
                         columns = c("ID", "Variable Name", "Unit"))

}
