% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnet_calc_mp.R
\name{mnet_calc_mp}
\alias{mnet_calc_mp}
\title{Calculate soil matric potential for Oklahoma Mesonet data}
\usage{
mnet_calc_mp(data)
}
\arguments{
\item{data}{a data frame that contains columns for delta-T temperature change
data (i.e. TR05, TR25, TR60, TR75)}
}
\value{
a data frame containing new columns with matric potential (kPa) for each
 column of delta-T temperature change data.
}
\description{
Calculate soil matric potential from delta-T soil temperature change data
 from the Oklahoma Mesonet using the equation from Zhang et al (2019): MP = 2083/(1+ exp(-3.35*(Tref-3.17)));
 where Tref is the measured delta-T data
}
