% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnet_root_url.R
\name{mnet_root_url}
\alias{mnet_root_url}
\title{Root url for Oklahoma Mesonet Time Series files}
\usage{
mnet_root_url(data_source = "mesonet")
}
\arguments{
\item{data_source}{the data source for which to provide the url: "mesonet"
for the Oklahoma Mesonet, "fcars" for the USDA ARS station network in the
Fort Cobb watershed, or "ars" for the USDA ARS station network in the
Little Washita watershed}
}
\description{
Root url for Oklahoma Mesonet Time Series files
}
